+++ src/Sql/Custom.php
@@ -1,8 +1,8 @@
 <?php
 
 namespace Katrina\Sql;
+
 use Katrina\Connection\DB as DB;
-use Katrina\Exception\Exception;
 use Katrina\Sql\Pagination as Pagination;
 use PDO;
 
@@ -19,10 +19,10 @@ abstract class Custom extends Pagination
             $stmt = DB::query($query);
             $stmt->execute();
             $res = $stmt->fetchAll(PDO::FETCH_ASSOC);
-        
+
             return $res;
         } catch (\PDOException $e) {
-            Exception::alertMessage($e, "'customQueryAll()' error");
+            throw new \PDOException("Error in 'customQueryAll()': " . $e->getMessage());
         }
     }
 
@@ -37,10 +37,10 @@ abstract class Custom extends Pagination
             $stmt = DB::query($query);
             $stmt->execute();
             $res = $stmt->fetch(PDO::FETCH_ASSOC);
-        
+
             return $res;
         } catch (\PDOException $e) {
-            Exception::alertMessage($e, "'customQueryOnly()' error");
+            throw new \PDOException("Error in 'customQueryOnly()': " . $e->getMessage());
         }
     }
 }