+++ src/Sql/TypesTrait.php
@@ -2,12 +2,17 @@
 
 namespace Katrina\Sql;
 
-abstract class Types
+trait TypesTrait
 {
     /**
-     * @return Types
+     * @var string
      */
-    public function primary(): Types
+    private string $growing = "ASC";
+    
+    /**
+     * @return self
+     */
+    public function primary(): self
     {
         $this->sql = rtrim($this->sql, ",");
         $this->sql .= " PRIMARY KEY ";
@@ -16,9 +21,9 @@ abstract class Types
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function notNull(): Types
+    public function notNull(): self
     {
         $this->sql = rtrim($this->sql, ",");
         $this->sql .= " NOT NULL,";
@@ -29,19 +34,19 @@ abstract class Types
     /**
      * @param string $table
      * 
-     * @return Types
+     * @return self
      */
-    public function alter(string $table): Types
+    public function alter(string $table): self
     {
         $this->sql = "ALTER TABLE `$table` ";
 
         return $this;
     }
-    
+
     /**
-     * @return Types
+     * @return self
      */
-    public function add(): Types
+    public function add(): self
     {
         $this->sql .= "ADD COLUMN ";
 
@@ -49,9 +54,9 @@ abstract class Types
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function modify(): Types
+    public function modify(): self
     {
         $this->sql .= "MODIFY COLUMN ";
 
@@ -61,9 +66,9 @@ abstract class Types
     /**
      * @param string $old_column
      * 
-     * @return Types
+     * @return self
      */
-    public function change(string $old_column): Types
+    public function change(string $old_column): self
     {
         $this->sql .= "CHANGE COLUMN `$old_column` ";
 
@@ -74,9 +79,9 @@ abstract class Types
      * @param string $old_table
      * @param string $new_name
      * 
-     * @return Types
+     * @return self
      */
-    public function rename(string $old_table, string $new_name): Types
+    public function rename(string $old_table, string $new_name): self
     {
         $this->sql .= "RENAME TABLE `$old_table` TO `$new_name`,";
 
@@ -86,9 +91,9 @@ abstract class Types
     /**
      * @param string $column
      * 
-     * @return Types
+     * @return self
      */
-    public function drop(string $column): Types
+    public function drop(string $column): self
     {
         $this->sql .= "DROP COLUMN $column;";
 
@@ -98,10 +103,10 @@ abstract class Types
     /**
      * @param string $foreign_key
      * 
-     * @return Types
+     * @return self
      */
-    public function foreign(string $foreign_key): Types
-    {   
+    public function foreign(string $foreign_key): self
+    {
         $this->sql .= "FOREIGN KEY (`$foreign_key`) ";
 
         return $this;
@@ -110,9 +115,9 @@ abstract class Types
     /**
      * @param string $constraint
      * 
-     * @return Types
+     * @return self
      */
-    public function constraint(string $constraint): Types
+    public function constraint(string $constraint): self
     {
         $this->sql .= "CONSTRAINT `$constraint` ";
 
@@ -122,22 +127,22 @@ abstract class Types
     /**
      * @param string $constraint
      * 
-     * @return Types
+     * @return self
      */
-    public function addConstraint(string $constraint): Types
+    public function addConstraint(string $constraint): self
     {
         $this->sql .= "ADD CONSTRAINT `$constraint` ";
 
         return $this;
     }
-    
+
     /**
      * @param string $references
      * @param string $id
      * 
-     * @return Types
+     * @return self
      */
-    public function references(string $references, string $id): Types
+    public function references(string $references, string $id): self
     {
         $this->sql .= "REFERENCES `$references`(`$id`),";
 
@@ -147,14 +152,14 @@ abstract class Types
     /**
      * @param string $default
      * 
-     * @return Types
+     * @return self
      */
-    public function default(string $default): Types
+    public function default(string $default): self
     {
         $this->sql = rtrim($this->sql, ",");
-        
+
         if ($default == 'CURRENT_TIMESTAMP' || $default == 'current_timestamp') {
-            $default = strtoupper($default);    
+            $default = strtoupper($default);
             $this->sql .= " DEFAULT $default,";
         } else {
             $this->sql .= " DEFAULT '$default',";
@@ -164,20 +169,20 @@ abstract class Types
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function unique(): Types
+    public function unique(): self
     {
-        $this->sql = rtrim($this->sql, ",");    
+        $this->sql = rtrim($this->sql, ",");
         $this->sql .= " UNIQUE,";
 
         return $this;
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function unsigned(): Types
+    public function unsigned(): self
     {
         $this->sql = rtrim($this->sql, ",");
         $this->sql .= " UNSIGNED,";
@@ -188,9 +193,9 @@ abstract class Types
     /**
      * @param string $column
      * 
-     * @return Types
+     * @return self
      */
-    public function after(string $column): Types
+    public function after(string $column): self
     {
         $comma = substr($this->sql, -1);
 
@@ -204,9 +209,9 @@ abstract class Types
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function first(): Types
+    public function first(): self
     {
         $comma = substr($this->sql, -1);
 
@@ -220,25 +225,25 @@ abstract class Types
     }
 
     /**
-     * @return Types
+     * @return self
      */
-    public function increment(): Types
+    public function increment(): self
     {
         $this->sql = rtrim($this->sql, ",");
         $this->sql .= " AUTO_INCREMENT,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function boolean(string $field): Types
+    public function boolean(string $field): self
     {
         $this->sql .= "`$field` BOOLEAN,";
-        
+
         return $this;
     }
 
@@ -247,12 +252,12 @@ abstract class Types
      * @param int $value1
      * @param int $value2
      * 
-     * @return Types
+     * @return self
      */
-    public function decimal(string $field, int $value1, int $value2): Types
+    public function decimal(string $field, int $value1, int $value2): self
     {
         $this->sql .= "`$field` DECIMAL($value1, $value2),";
-        
+
         return $this;
     }
 
@@ -260,73 +265,73 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function char(string $field, int $size): Types
+    public function char(string $field, int $size): self
     {
         $this->sql .= "`$field` CHAR($size),";
-        
+
         return $this;
     }
-    
+
     /**
      * @param string $field
      * @param string $size
      * 
-     * @return Types
+     * @return self
      */
-    public function varchar(string $field, int $size): Types
+    public function varchar(string $field, int $size): self
     {
-        $this->sql .= "`$field` VARCHAR($size),";   
-        
+        $this->sql .= "`$field` VARCHAR($size),";
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function tinytext(string $field): Types
+    public function tinytext(string $field): self
     {
         $this->sql .= "`$field` TINYTEXT,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function mediumtext(string $field): Types
+    public function mediumtext(string $field): self
     {
         $this->sql .= "`$field` MEDIUMTEXT,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function longtext(string $field): Types
+    public function longtext(string $field): self
     {
         $this->sql .= "`$field` LONGTEXT,";
-        
+
         return $this;
     }
-    
+
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function text(string $field): Types
+    public function text(string $field): self
     {
         $this->sql .= "`$field` TEXT,";
-        
+
         return $this;
     }
 
@@ -334,12 +339,12 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function tinyint(string $field, int $size): Types
+    public function tinyint(string $field, int $size): self
     {
         $this->sql .= "`$field` TINYINT($size),";
-        
+
         return $this;
     }
 
@@ -347,12 +352,12 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function smallint(string $field, int $size): Types
+    public function smallint(string $field, int $size): self
     {
         $this->sql .= "`$field` SMALLINT($size),";
-        
+
         return $this;
     }
 
@@ -360,12 +365,12 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function mediumint(string $field, int $size): Types
+    public function mediumint(string $field, int $size): self
     {
         $this->sql .= "`$field` MEDIUMINT($size),";
-        
+
         return $this;
     }
 
@@ -373,12 +378,12 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function bigint(string $field, int $size): Types
+    public function bigint(string $field, int $size): self
     {
         $this->sql .= "`$field` BIGINT($size),";
-        
+
         return $this;
     }
 
@@ -386,72 +391,137 @@ abstract class Types
      * @param string $field
      * @param int $size
      * 
-     * @return Types
+     * @return self
      */
-    public function int(string $field, int $size = 11): Types
+    public function int(string $field, int $size = 11): self
     {
         $this->sql .= "`$field` INT($size),";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function date(string $field): Types
+    public function date(string $field): self
     {
         $this->sql .= "`$field` DATE,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function year(string $field): Types
+    public function year(string $field): self
     {
         $this->sql .= "`$field` YEAR,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function time(string $field): Types
+    public function time(string $field): self
     {
         $this->sql .= "`$field` TIME,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function datetime(string $field): Types
+    public function datetime(string $field): self
     {
         $this->sql .= "`$field` DATETIME,";
-        
+
         return $this;
     }
 
     /**
      * @param string $field
      * 
-     * @return Types
+     * @return self
      */
-    public function timestamp(string $field): Types
+    public function timestamp(string $field): self
     {
         $this->sql .= "`$field` TIMESTAMP,";
-        
+
+        return $this;
+    }
+
+    /**
+     * @param int $offset
+     * @param int $row_count
+     * 
+     * @return self
+     */
+    public function limit(int $offset, int $row_count): self
+    {
+        if (!is_int($offset)) {
+            throw new \IntlException("Offset not INT");
+        } else if (!is_int($row_count)) {
+            throw new \IntlException("Row count not INT");
+        }
+
+        $this->sql = rtrim($this->sql, ",");
+        $this->sql .= " LIMIT $offset, $row_count";
+
+        return $this;
+    }
+
+    /**
+     * @param string $like
+     * 
+     * @return self
+     */
+    public function like(string $like): self
+    {
+        $this->sql = rtrim($this->sql, ",");
+        $this->sql .= " LIKE '" . $like . "'";
+
+        return $this;
+    }
+
+    /**
+     * @param string $column
+     * @param bool $asc
+     * 
+     * @return self
+     */
+    public function order(string $column, bool $asc = true): self
+    {
+        if ($asc == false) {
+            $this->growing = "DESC";
+        }
+
+        $this->sql = rtrim($this->sql, ",");
+        $this->sql .= " ORDER BY " . $column . " " . $this->growing;
+
+        return $this;
+    }
+
+    /**
+     * @param mixed $first_date
+     * @param mixed $second_date
+     * 
+     * @return self
+     */
+    public function between($first_date, $second_date): self
+    {
+        $this->sql = rtrim($this->sql, ",");
+        $this->sql .= " BETWEEN " . $first_date . " AND " . $second_date;
+
         return $this;
     }
 }